moved payment clearing to controller class to throw real timeout exceptions, added iteration code
