+++ src/paymentservicejava/src/main/java/com/hipstershop/paymentservicejava/dataaccess/PaymentRecordRepository.java
@@ -7,5 +7,5 @@ import com.hipstershop.paymentservicejava.model.PaymentRecord;
 
 @Repository
 public interface PaymentRecordRepository extends CrudRepository<PaymentRecord, Long>{
-    
+
 }
