+++ src/paymentservicejava/src/main/resources/application.properties
@@ -4,8 +4,8 @@ spring.datasource.url=jdbc:mariadb://localhost:3306/paymentdb
 spring.datasource.username=swagstore
 spring.datasource.password=weLoveSwagAtDash2023
 spring.datasource.driver-class-name=org.mariadb.jdbc.Driver
-# Test settings
 #spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MariaDBDialect
+# Test settings
 #spring.datasource.url=jdbc:h2:mem:paymentdb;DB_CLOSE_DELAY=-1;NON_KEYWORDS=KEY,VALUE
 #spring.datasource.driver-class-name=org.h2.Driver
 #spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.H2Dialect
